refactor: use pinia to init game
